use_cases/nprint_ids_case/requirements.txt
autogluon-tabular==0.1.0 ; python_version >= "3.7" and python_version < "3.9"
lightgbm==3.3.2 ; python_version >= "3.7" and python_version < "3.9"
matplotlib==3.3.4 ; python_version >= "3.7" and python_version < "3.9"
nprintml==1.1.2 ; python_version >= "3.7" and python_version < "3.9"
numpy==1.19.5 ; python_version >= "3.7" and python_version < "3.9"
pandas==1.1.5 ; python_version >= "3.7" and python_version < "3.9"
scikit-learn==0.24.2 ; python_version >= "3.7" and python_version < "3.9"
scipy==1.5.4 ; python_version >= "3.7" and python_version < "3.9"
trustee==1.1.4 ; python_version >= "3.7" and python_version < "3.9"
